feat(door): #44 문닫힘 감지 버튼 실시간으로 반응하도록 수정
